+++ src/docs/pages/guide/config/index.vue
@@ -331,6 +331,9 @@
     - `change-on-select` : 选中即改变
     - `source-to-target-text` : 源至目标文案
     - `target-to-source-text` : 目标至源文案
+    - `palettes` : 色板
+    - `only-palettes` : 仅使用色板
+    - `inside-padding` : 内部填充
 
     </script>
     </doc-guide>