+++ src/lib/components/colorpicker/index.vue
@@ -10,6 +10,8 @@
         :clearable="clearable"
         :value-type="valueType"
         :allow-alpha="allowAlpha"
+        :palettes="palettes"
+        :only-palettes="onlyPalettes"
     >
     
     <div class="form-name" v-if="!conf.hideName && !!conf.formName">{{conf.formName}}</div>
@@ -43,10 +45,13 @@
         <div class="picker">
             <div
                 class="panel"
+                :class="{
+                        'not-allow' : onlyUsePalettes
+                    }"
                 :style="{
                     'background-color' : colorH
                 }"
-
+                   
                 @mousedown="_pickColor($event);_hslHSync(false)"
                 @mouseup="_hslHSync(true)"
             >
@@ -63,7 +68,7 @@
                     @mouseup.capture="_stopmoveStraw"
                 ></div>
             </div>
-            <div class="tools">
+            <div class="tools" v-show="!onlyUsePalettes">
                 <div
                     class="color-copy"
                      :id="'mor-colorpicker-copy-'+uiid"
@@ -115,7 +120,7 @@
                     </div>
                 </div>
             </div>
-            <div class="values">
+            <div class="values" v-show="!onlyUsePalettes">
                 <div class="input">
                     <div class="rgba" v-if="data.showValueType === 'rgba' && typeof data.colorValue === 'object'">
                         <morning-textinput v-model="data.colorValue.r" :state="inputIsReadonly ? 'readonly' : 'normal'" @focus="_hslHSync(true)" @blur="_rgbaChangeR();_hslHSync(false)"></morning-textinput>
@@ -152,11 +157,94 @@
                     <i class="mo-icon mo-icon-sort"></i>
                 </div>
             </div>
-           <!--  <div class="palettes">
-                <ul class="colors">
-                    <li></li>
-                </ul>
-            </div> -->
+
+            <template v-if="conf.palettes.length > 0 && typeof conf.palettes[0] === 'string'">
+                <div class="palettes">
+                    <template v-for="(color, index) in conf.palettes">
+                        <div
+                            class="palettes-color-item"
+                            :key="index"
+                            :style="{
+                                'background-color' : color
+                            }"
+                            :id="'mor-colorpicker-palettes-tip-'+uiid+'-'+index"
+                            :title="color"
+                            @click="set(color)"
+                        >
+                        </div>
+                        <morning-tip
+                            :key="index"
+                            :target="'#mor-colorpicker-palettes-tip-'+uiid+'-'+index"
+                            color="light-blue"
+                            offset="3px 0"
+                        >
+                            {{color}}
+                        </morning-tip>
+                    </template>
+                </div>
+            </template>
+
+            <template v-else-if="conf.palettes.length > 0 && typeof conf.palettes[0] === 'object'">
+                <div class="palettes multi-palettes">
+                    <template v-for="(color, index) in conf.palettes[data.currentPalette].colors">
+                        <div
+                            class="palettes-color-item"
+                            :key="index"
+                            :style="{
+                                'background-color' : color
+                            }"
+                            :id="'mor-colorpicker-palettes-tip-'+uiid+'-'+index"
+                            :title="color"
+                            @click="set(color)"
+                        >
+                        </div>
+                        <morning-tip
+                            :key="index"
+                            :target="'#mor-colorpicker-palettes-tip-'+uiid+'-'+index"
+                            color="light-blue"
+                            offset="3px 0">
+                            {{color}}
+                        </morning-tip>
+                    </template>
+                    <div
+                        class="toggle-palette"
+                        @click="_togglePalettesPicker"
+                    >
+                        <i class="mo-icon mo-icon-block-4"></i>
+                    </div>
+                    <div
+                        class="palettes-picker"
+                        :class="{show : data.palettesPicker}"
+                    >
+                        <div class="palette-title">
+                            <span>选择色板</span>
+                            <i class="mo-icon mo-icon-close" @click="_togglePalettesPicker"></i>
+                        </div>
+                        <div class="palette-item-list">
+                            <div
+                                v-for="(item, index) in conf.palettes"
+                                :key="index"
+                                class="palette-item"
+                                @click="_usePalettes(index);_togglePalettesPicker()"
+                            >
+                                <div class="palette-name">{{item.name}}</div>
+                                <div class="palette-color-preview">
+                                    <div
+                                        v-for="(previewColor, previewIndex) in item.colors.slice(0, 6)"
+                                        :key="previewIndex"
+                                        class="palettes-color-item"
+                                        :style="{
+                                            'background-color' : previewColor
+                                        }"
+                                    >
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </template>
+
         </div>
     </div>
 
@@ -199,6 +287,14 @@ export default {
         allowAlpha : {
             type : Boolean,
             default : true
+        },
+        palettes : {
+            type : Array,
+            default : (() => [])
+        },
+        onlyPalettes : {
+            type : Boolean,
+            default : false
         }
     },
     computed : {
@@ -206,7 +302,9 @@ export default {
 
             return {
                 valueType : this.valueType,
-                allowAlpha : this.allowAlpha
+                allowAlpha : this.allowAlpha,
+                palettes : this.palettes,
+                onlyPalettes : this.onlyPalettes
             };
 
         },
@@ -257,6 +355,11 @@ export default {
 
             return Math.round(this.data.alpha / maxAlpha * num100) / num100;
 
+        },
+        onlyUsePalettes : function () {
+
+            return this.conf.onlyPalettes && this.conf.palettes.length > 0;
+
         }
     },
     data : function () {
@@ -290,7 +393,9 @@ export default {
                 hslHReversal : null,
                 colorValue : null,
                 copyNote : null,
-                copyNoteTimeout : null
+                copyNoteTimeout : null,
+                currentPalette : 0,
+                palettesPicker : false
             }
         };
 
@@ -643,12 +748,24 @@ export default {
 
         },
         _moveStraw : function () {
+
+            if (this.onlyUsePalettes) {
+
+                return;
+
+            }
           
             this.data.dontPickColor = true;
             this._moveItemRecord(0);
 
         },
         _stopmoveStraw : function () {
+
+            if (this.onlyUsePalettes) {
+
+                return;
+
+            }
 
             this.data.dontPickColor = false;
             this._moveMouseup();
@@ -667,6 +784,12 @@ export default {
 
         },
         _pickColor : function (evt) {
+
+            if (this.onlyUsePalettes) {
+
+                return;
+
+            }
 
             if (this.data.dontPickColor ||
                 this.conf.state === 'disabled' ||
@@ -848,6 +971,16 @@ export default {
                 this.Vue.nextTick(() => this.$refs[`mor-colorpicker-copytip-${this.uiid}`].position());
 
             }, copyShowtime);
+
+        },
+        _togglePalettesPicker : function () {
+
+            this.data.palettesPicker = !this.data.palettesPicker;
+
+        },
+        _usePalettes : function (index) {
+
+            this.data.currentPalette = index;
 
         },
         set : function (value) {